Fix for extra indent on multi-line descriptions
The """ inserted at the end of the docstring was causing problems.
We don't need to find the end of the paragraph of the last insertion,
since we already have a mark at the end of the region.

Also made number of spaces indented an optional parameter, so that the
long description isn't indented if it has multiple lines.